agentserver/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
pytest==8.3.3
